requirements.txt
telethon>=1.34.0
aiogram>=3.3.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
supabase>=2.0.0
cryptography>=42.0.0

